Embed maven-scm-api and -manager rather than use mojo-executor
This frees us from some of the bugs of the mojo-executor with tighter control
of the scm lifecycle. We also shave a bit of time spent setting up and executing
the mojo. It also allows tests to be written much easier.

This commit also fixes a bug where it was impossible to access releases in
tags with an SCM which did not represent the release information in the
scm url, i.e. subversion tag information versus git.