Start work on removing git from the build strategy and integrate it with the mojo
This is necessary in order to reduce duplication of having each strategy perform
their own git operations. This also reduces the responsibility of a build strategy
to source code retrieval only. The only byproduct is that a build strategy now
does not have control over how source code is built. Rather, any available strategy
can now be considered.

This is a partial commit which breaks existing APIs. Some files are yet to be
updated with changes.