Move artifact installation into the renamed dependency collector
This also fixes a bug where by it may have been possible for ignored
artifacts to be built. This would generally happen as part of
dependencies of ignored artifacts which were also dependencies of
an artifact that we were going to build.

This is an effort to refractor the build dependencies class to remove
its god status.